♻️ merge github projectors
