+++ backend/event-listeners/src/domain/projections/github_users.rs
@@ -1,9 +1,10 @@
 use ::infrastructure::database::schema::*;
+use derive_getters::Getters;
 use derive_new::new;
 use domain::GithubUserId;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Clone, Insertable, AsChangeset, Serialize, Deserialize, new)]
+#[derive(Debug, Clone, Insertable, AsChangeset, Serialize, Deserialize, new, Getters)]
 pub struct GithubUser {
 	id: GithubUserId,
 	login: String,
