--- backend/event-listeners/src/domain/projectors/github/mod.rs
@@ -1,4 +1,0 @@
-pub mod contributors;
-pub mod crm;
-pub mod issues;
-pub mod users;
