+++ backend/event-listeners/src/domain/projectors/mod.rs
@@ -3,8 +3,5 @@ mod github;
 mod project;
 
 pub use budget::Projector as BudgetProjector;
-pub use github::{
-	contributors::Projector as GithubNewContributorsProjector, crm::Projector as CrmProjector,
-	issues::Projector as GithubIssuesRepositoryProjector, users::Projector as GithubUsersProjector,
-};
+pub use github::Projector as GithubProjector;
 pub use project::Projector as ProjectProjector;
